# docs/openapi/gateway.yaml
openapi: 3.0.3
info:
  title: API Gateway
  version: "1.0.0"
  description: |
    Маршрутизирует запросы между фронтом и микросервисами системы бронирования.
servers:
  - url: http://localhost:8000
paths:
...
  /admin/zones:
    post:
      summary: Администрирование зон (создание/редактирование/удаление)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ZoneAdminRequest' }
      responses:
        '200': { description: ОК }
        '403': { description: Нет доступа }
  /admin/zones/{id}/close:
    post:
      summary: Закрыть зону на обслуживание
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ZoneCloseRequest' }
      responses:
        '200': { description: ОК }
        '403': { description: Нет доступа }
  /notify/email:
    post:
      summary: Отправить email-уведомление пользователю
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailNotification' }
      responses:
        '200': { description: Email отправлен }
        '400': { description: Ошибка }
  /notify/bulk:
    post:
      summary: Массовая email-рассылка
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BulkNotification' }
      responses:
        '200': { description: Email отправлены }
        '400': { description: Ошибка }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ZoneAdminRequest:
      type: object
      properties:
        name: { type: string }
        address: { type: string }
    ZoneCloseRequest:
      type: object
      properties:
        reason: { type: string }
        maintenance_period: { type: string }
    EmailNotification:
      type: object
      properties:
        email: { type: string }
        subject: { type: string }
        text: { type: string }
    BulkNotification:
      type: object
      properties:
        emails:
          type: array
          items: { type: string }
        subject: { type: string }
        text: { type: string }